pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "python"]
build-backend = "setuptools.build_meta"

[project]
name = "algebratic-graph-rag"
version = "0.1.0"
description = "PageRank-based Hierarchical GraphRAG System with Academic Evaluation Framework"
dependencies = [
    "python-dotenv",
    "neo4j",
    "openai",
    "sentence-transformers",
    "networkx",
    "numpy",
    "scikit-learn",
    "langchain",
    "langchain-openai", 
    "langchain-community",
    "pypdf",
    "pydantic",
    "typing-extensions",
    
    # Academic Evaluation Dependencies
    "pandas>=1.5.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "scipy>=1.9.0",
    "jupyter",
    
    # Optional visualization
    "plotly",
    "bokeh",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]

evaluation = [
    "pandas>=1.5.0",
    "matplotlib>=3.5.0", 
    "seaborn>=0.11.0",
    "scipy>=1.9.0",
    "jupyter",
    "plotly",
    "bokeh",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
